enable user configuration of curl options connecttimeout and timeout
The values of these new keys are set to the defaults of curl.
Now the user can also change these two values for their requirements.
When server errors occur on amazons side (which can happen, as on dec 7th) you
can now configure to just let the user wait for example 60s before it times out
instead of 300s on each request/process.